% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{a}
\alias{a}
\title{Example results resample()}
\format{
A list with 9 items, returned by \code{\link[=resample]{resample()}}.
}
\usage{
a
}
\description{
Results of a fitted generalised linear model, returned by \code{\link[=resample]{resample()}}. Fitted to example Rhyssinae wasp data with 99 resamples and pairwise comparisons between forest types. Used in several of the examples to save time, since it can take several minutes to run the analyses.
}
\keyword{datasets}
