% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{as.datetime}
\alias{as.datetime}
\title{Convert to datetime}
\usage{
as.datetime(x)
}
\arguments{
\item{x}{Character vector giving dates, times and time zone offsets. Should ideally be in the format "2014-10-14 11:04:00 UTC+03:00", but may convert OK even if e.g. the time is missing. Can also be a datetime object, in which case \code{x} is returned without changes.}
}
\value{
Vector of datetime objects.
}
\description{
Convert strings of the format "2014-10-14 11:04:00 UTC+03:00" to datetime objects.
}
\examples{

# save two dates as string
x = c("2014-10-14 11:04:00 UTC+03:00", "2014-11-02 01:04:00 UTC-02:30")

# convert to datetime
x = as.datetime(x)
 
# try out different operations with datetimes
x
x + 15
x[2] - x[1]
floor(x)
c(x, x)
seq(x[1], x[2], by=3600*24)

# datetimes are converted to character when saved in a data frame
m = data.frame(a=1:2, x=1:2)
m$x = x
m

}
\seealso{
\code{\link[=get_date]{get_date()}}, \code{\link[=is_datetime]{is_datetime()}} and \code{\link[=set_tz]{set_tz()}}. The following base functions have been modified to work with datetimes:
\itemize{
\item \code{\link[=as.character.datetime]{as.character.datetime()}}, \code{\link[=c.datetime]{c.datetime()}}, \code{\link[=ceiling.datetime]{ceiling.datetime()}}, \code{\link[=floor.datetime]{floor.datetime()}}, \code{\link[=length.datetime]{length.datetime()}}, \code{\link[=print.datetime]{print.datetime()}}, \code{\link[=seq.datetime]{seq.datetime()}}
\item basic operators such as \code{+}, \code{-} (see examples)
\item data frames accept datetimes (but convert them to character)
}
}
