% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{as.interval}
\alias{as.interval}
\title{Convert to interval}
\usage{
as.interval(s, e)
}
\arguments{
\item{s}{Vector of datetimes giving the start of each interval. Either datetime objects (see \code{\link[=as.datetime]{as.datetime()}}) or something that can be converted to datetime.}

\item{e}{Vector of datetimes giving the end of each interval. Either datetime objects (see \code{\link[=as.datetime]{as.datetime()}}) or something that can be converted to datetime.}
}
\value{
Vector of interval objects.
}
\description{
Convert vectors giving the start and end of time intervals to interval objects.
}
\examples{

# save start and end datetimes
s = as.datetime(c("2014-01-14 00:00:00 UTC+00:00", "2014-01-14 12:00:00 UTC+00:00"))
e = as.datetime(c("2014-01-15 00:00:00 UTC+00:00", "2014-01-15 12:00:00 UTC+00:00"))
 
# save as interval
x = as.interval(s, e)

# try out different operations with intervals
x
x + 15
c(x, x[1])

# get length in seconds, and overlap
tdiff(x)
overlap(x, x[1])

}
\seealso{
\code{\link[=as.datetime]{as.datetime()}}, \code{\link[=is_interval]{is_interval()}}, \code{\link[=means]{means()}}, \code{\link[=overlap]{overlap()}}, \code{\link[=overlaps]{overlaps()}}, \code{\link[=tdiff]{tdiff()}}. The following base functions have been modified to work with datetimes:
\itemize{
\item \code{\link[=c.interval]{c.interval()}}, \code{\link[=length.interval]{length.interval()}}, \code{\link[=max.interval]{max.interval()}}, \code{\link[=min.interval]{min.interval()}}, \code{\link[=print.interval]{print.interval()}}
\item basic operators such as \code{+}, \code{-} (see examples)
}
}
