% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{as_h_min}
\alias{as_h_min}
\title{Get time zone offsets of datetimes as hours and minutes}
\usage{
as_h_min(secs)
}
\arguments{
\item{secs}{Vector of time zone offsets in seconds.}
}
\value{
Character vector of time zone offsets in hours and minutes.
}
\description{
Helper function for getting the time zone offsets of datetime objects in the format "03:00". Used by e.g. \code{\link[=as.character.datetime]{as.character.datetime()}}.
}
