% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{as_italic}
\alias{as_italic}
\title{Print text as italic}
\usage{
as_italic(x)
}
\arguments{
\item{x}{Character vector which is to be converted to italic.}
}
\value{
Vector which when displayed in a plot, is in italics.
}
\description{
Make a character vector (typically species names) display in italic font. Currently makes all the text italics. To mix italic and non-italic font, you will need to mess about with stuff like e.g. \code{expression("Species one "*"n=199")} or function \code{\link[=bquote]{bquote()}}.
}
