% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_datetime.R
\name{axis_datetime}
\alias{axis_datetime}
\title{Label the x axis of datetime plots}
\usage{
axis_datetime(
  xlim,
  ticks = TRUE,
  format = "\%Y-\%m-\%d",
  breaks = default_breaks(),
  ...
)
}
\arguments{
\item{xlim}{Interval object (see \code{\link[=as.interval]{as.interval()}} giving the left and right limits of the plot.}

\item{ticks}{If TRUE (the default), draw tick marks above the label. These tick marks cannot be customised by the user: if wanting custom line widths, colours etc, you will have to draw them separately.}

\item{format}{Character string telling how to format the labels. See \code{\link[=strftime]{strftime()}} for accepted formats. Default is year-month-day.}

\item{breaks}{Vector of datetime objects (see \code{\link[=as.datetime]{as.datetime()}}) telling where to place labels. Anything outside the plot area is ignored. If not given, uses default breaks provided by \code{\link[=default_breaks]{default_breaks()}}, which are at about three month intervals.}

\item{...}{Other arguments passed to \code{\link[=text]{text()}} (which handles the actual drawing of the labels). Typically these can include \code{srt} to adjust the angle at which labels are drawn, and \code{cex} to adjust the label size.}
}
\value{
List of the arguments given to \code{\link[=text]{text()}}, returned invisibly. These may be useful as a starting point if wanting to create custom labels, especially arguments \code{x} and \code{y} which give the plot coordinates of each label.
}
\description{
Add labels and tick marks to the x axis of plots that have datetimes on the x axis, such as plots created by \code{\link[=plot_time]{plot_time()}}. Called by \code{\link[=plot_time]{plot_time()}} by default.
}
