% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{c.datetime}
\alias{c.datetime}
\title{Concatenate datetimes}
\usage{
\method{c}{datetime}(...)
}
\arguments{
\item{...}{Vectors of datetimes.}
}
\value{
Vector of datetimes.
}
\description{
Join datetime objects with \code{\link[=c]{c()}}.
}
