% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{c.interval}
\alias{c.interval}
\title{Concatenate intervals}
\usage{
\method{c}{interval}(...)
}
\arguments{
\item{...}{Vectors of intervals.}
}
\value{
Vector of intervals.
}
\description{
Join interval objects with \code{\link[=c]{c()}}.
}
