% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{combine_columns}
\alias{combine_columns}
\title{Combine columns}
\usage{
combine_columns(x, what, sep = " ", all = FALSE)
}
\arguments{
\item{x}{Data frame. Usually wasp data where each row is an individual wasp.}

\item{what}{Character vector giving the names of the columns of \code{x} that are to be combined.}

\item{sep}{String to separate the values of the columns with. Default is a space: e.g. "primary" and "Uganda 2014-2015" become "primary Uganda 2014-2015".}

\item{all}{If FALSE (the default), simply combines the values in the columns. If TRUE, combines the values and gives the result as a factor, whose levels include all possible combinations of column values. See 'Details'.}
}
\value{
Character vector or factor of the same length as there are rows in \code{x}. Will be a character vector if \code{all} is FALSE and none of the combined columns were factors. Otherwise a factor.
}
\description{
Combine the columns in a data frame. Mainly for combining e.g. the forest type and collecting event of wasp data for easy use in rarefaction plots etc. A handier wrapper for \code{\link[=paste]{paste()}}, since it preserves factor levels and their order.
}
\details{
The function tries to preserve the order of any factor levels. So if any of the columns is a factor, it will return a factor whose levels are in the same order. (e.g. if "primary" was the first forest type, then "primary Uganda 2014-2015" will come before "disturbed Uganda 2014-2015".)

If \code{all} is TRUE, the function will return a factor whose levels include all possible combinations of the combined columns. I.e. if combining forest type (5 levels) and event (5 levels), there will be 5*5=25 levels, even if the data only consisted of one row which combined to "primary Uganda 2014-2015".

If 'all' is FALSE and none of the columns is a factor, returns a character vector.
}
\examples{
# make a simple example dataset
x = data.frame(type=c("disturbed", "primary"), event=c("Uganda", "Amazon"))
x

# combine
combine_columns(x, c("type", "event"))

# combine and show the four possible factor levels
combine_columns(x, c("type", "event"), all=TRUE)

# make the first column a factor with three levels
x$type = factor(x$type, levels=c("primary", "disturbed", "farm"))

# combine and show the six possible factor levels
combine_columns(x, c("type", "event"), sep="_", all=TRUE)

}
