% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{default_breaks}
\alias{default_breaks}
\title{Get default breaks for label_plot_time()}
\usage{
default_breaks()
}
\value{
Vector of datetime objects.
}
\description{
Helper function used by \code{\link[=axis_datetime]{axis_datetime()}}. Gets the default breaks for x axis tick marks. These are generally every three months, during the five collecting events.
}
\keyword{internal}
