% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_legend.R
\name{default_legend}
\alias{default_legend}
\title{Default legend}
\usage{
default_legend(what = "forest_type", event = NULL, ...)
}
\arguments{
\item{what}{What legend to add. Can be "forest type", in which case shows the forest types. Or can be a character vector of taxon names, in which case shows the taxa. Taxa are got from the factor levels of the vector, so it is OK to have the same taxon several times.}

\item{event}{Character vector of what collecting events to include if making a forest type legend. Accepted events are "Uganda 2014-2015", "Amazon 1998", "Amazon 2000", "Amazon 2008", "Amazon 2011". \emph{Alternatively}, can be a character vector of locations (e.g. traps), in which case all collecting events involving those locations will be included. If NULL, all events are included.}

\item{...}{Other arguments passed to \code{\link[=legend]{legend()}}. Typically \code{x}, which gives the position of the legend (default is "topright"). Arguments \code{legend} and \code{fill} have no effect.}
}
\value{
List returned by \code{\link[=legend]{legend()}}. Returned invisibly.
}
\description{
Add default legend to plots. Either a legend showing the forest types (in default order with default colours), or taxa (with default colour scheme).
}
\examples{
# get example wasp data
f = system.file("extdata", "wasps_example.csv", package = "turkuwasps", mustWork = TRUE)
wasps = read_wasps(f)

# remove damaged samples and their wasps
tmp = ecology_usable(wasps)
x = tmp$wasps
m = tmp$samples

# plot traps
tmp = plot_place(x$trap, m)

# add legend
default_legend("forest_type", event=names(tmp)) 

}
