% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{default_order}
\alias{default_order}
\title{Relevel locations to default order}
\usage{
default_order(x)
}
\arguments{
\item{x}{Vector of traps, forest types or other locations. Either as strings or factor.}
}
\value{
Vector of locations, factored so that the factor levels are in default order.
}
\description{
Helper function. Gives defaults for what order to place traps, forest types or other locations. Factors a vector of locations so the factor levels are in default order.
}
\keyword{internal}
