% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{default_xlims}
\alias{default_xlims}
\title{Get default x limits for plot_time()}
\usage{
default_xlims(x)
}
\arguments{
\item{x}{Vector of interval objects, telling when each wasp was collected.}
}
\value{
Interval object.
}
\description{
Helper function used by \code{\link[=plot_time]{plot_time()}}. Gets the default x limits to use for the plot.
}
