% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{draw_rarefaction}
\alias{draw_rarefaction}
\title{Draw rarefaction curves}
\usage{
draw_rarefaction(r, p = 0.95, add = FALSE, pch = NULL, pch_col = NULL, ...)
}
\arguments{
\item{r}{List of coordinates and standard errors for the curve returned by \code{\link[=get_rarefaction]{get_rarefaction()}}. Should have the x and y coordinates of the curve (\code{x}, \code{y}), and the standard errors (\code{se}). If the standard errors are not NULL, confidence intervals are drawn as well as the main curve.}

\item{p}{How large confidence intervals to draw around the rarefaction curves. Number between 0 and 1. Default (0.95) is to draw 95\% intervals, i.e. ± 1.96 standard errors. Used to estimate if two rarefaction curves are significantly different (e.g. for a significance of 0.05, check if the confidence intervals of two 95\% curves overlap).}

\item{add}{If TRUE, the rarefaction curve is added to an existing plot. Default is to create a new plot.}

\item{pch}{What symbols to use on the curve. Typically an integer between 0:18. See \code{\link[=points]{points()}} for accepted values. Default is for the curve to be drawn without symbols.}

\item{pch_col}{Colour to be used for the symbols.}

\item{...}{Graphical parameters passed to the two functions which draw the curves, \code{\link[=plot]{plot()}} and \code{\link[=lines]{lines()}}. These will override any default values such as colours. A few parameters (such as 'type' and 'pch') may not work as expected.}
}
\description{
Helper function used by \code{\link[=plot_rarefaction]{plot_rarefaction()}}. Draws rarefaction curves created by \code{\link[=get_rarefaction]{get_rarefaction()}}.
}
\keyword{internal}
