% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore how rain, temperature etc affected with wasp catches}
\usage{
explore(x, m, what = NULL, scale = "tdiff", res = 15, ...)
}
\arguments{
\item{x}{Vector of which sample each wasp came from. Chararacter string or factor.}

\item{m}{Data frame with the Malaise sample data. Must contain columns "name" and "event", and all the columns given by \code{what} and \code{scale}.}

\item{what}{Character vector giving the name of the columns of \code{m} that are to be plotted. If NULL, plots all numeric columns (e.g. rain, temperature etc) except for the column given by \code{scale}.}

\item{scale}{Character string giving the name of the column to use for scaling wasp catches. Default is to show number of wasps per trap day on the y axis. If empty (scale="") or the column isn't in \code{m}, shows number of wasps caught, irrespective of whether the sample was collected during one day or two weeks.}

\item{res}{Resolution of the grid. Default is to split the x and y axes into 15 grid lines, giving 14*14=196 grid cells.}

\item{...}{Other arguments passed to \code{\link[=image]{image()}}, which draws the grid cell colours. Will overwrite any default arguments such as \code{col}. (Argument \code{x} of \code{\link[=image]{image()}} cannot be overwritten)}
}
\description{
Plot extended scatterplots of wasp catches versus rain, temperature etc. Plot a scatterplot of e.g. rain on the x axis and number of wasps caught in each sample on the y axis. Then add:
\itemize{
\item A fitted line showing trends (e.g. decreasing catch the more rain there is).
\item A grid coloured in according to how many points there are in each grid cell (i.e. a 2D histogram). The counts are scaled columnwise, see "Notes".
}
}
\note{
The number of points in each grid cell is scaled columnwise before plotting. This means that the colour of each grid cell shows the \emph{portion} of points (of the column of cells) that are in the cell, not the absolute number of points in the cell.

E.g. if rainfall (in mm) is on the x axis, then each cell in column 5 mm - 10 mm shows what portion of the points between rain=5mm and rain=10mm are in the cell.

Without scaling like this, it gets hard to compare different levels of rainfall. The sample sizes (number of points) can be quite different in different columns.
}
\examples{
# get path to example wasp data
f = system.file("extdata", "wasps_example.csv", package = "turkuwasps", mustWork = TRUE)

# read the wasp data and get the corresponding sample data
tmp = read_wasps(f)
x = tmp$x
m = tmp$m

# show how catches varied with rain and amount of wood 
explore(x$sample, m, what=c("rain", "deadwood", "livewood"))

}
