% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter samples by collecting event of wasps}
\usage{
filter_samples(x, m)
}
\arguments{
\item{x}{Vector of which sample each wasp came from. Character or factor.}

\item{m}{Data frame with the Malaise sample data. Must contain columns "name" and "event".}
}
\value{
Data frame with the Malaise sample data. Only contains samples from the same collecting event(s) as the wasps came from.
}
\description{
Helper function used by \code{\link[=resample]{resample()}}. Filters samples based on the caught wasps. Only samples from the same collecting event(s) as the wasps are kept.
}
