% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{floor.datetime}
\alias{floor.datetime}
\title{Set the time of datetimes to the start of the day}
\usage{
\method{floor}{datetime}(x)
}
\arguments{
\item{x}{Vector of datetimes.}
}
\value{
Vector of converted datetimes.
}
\description{
Convert e.g. "2014-10-14 11:04:00 UTC+03:00" to "2014-10-14 00:00:00 UTC+03:00".
}
