% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_package_and_data.R
\docType{data}
\name{forest_type}
\alias{forest_type}
\title{Forest types}
\format{
A data frame with 3 variables. The variables are:
\itemize{
\item name Name of the forest type, e.g. "primary".
\item event Name of collecting event. One of "Uganda 2014-2015", "Amazon 1998", "Amazon 2000", "Amazon 2008", "Amazon 2011".
\item colour Colour to use by default when plotting.
}

Forest types are in the following order: Uganda, then Peru 1998, 2000, 2008, 2011. Within each grouping they are sorted by successional status (e.g. Uganda primary forest to farm).
}
\usage{
forest_type
}
\description{
Ugandan and Peruvian forest types. Used by some functions to get all the forest types, what order to plot them in by default, and what colours to use.
}
\keyword{datasets}
