% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{get_date}
\alias{get_date}
\title{Format a datetime as string}
\usage{
get_date(x, f = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{Vector of datetimes.}

\item{f}{String giving the desired format.}
}
\value{
Character vector.
}
\description{
Get datetimes as string in different formats. Default is to get the date, e.g. "2014-10-14". Accepts all formats used by \code{\link[=strptime]{strptime()}}, see \code{?strptime} for formats. Ugandan datetimes are in format "\%Y-\%m-\%d \%H:\%M:\%S UTC+03:00".
}
