% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_defaults}
\alias{get_defaults}
\title{Get default bar order and bar colours for barplots}
\usage{
get_defaults(x)
}
\arguments{
\item{x}{Vector of which trap, forest type or other location each wasp came from. Either as strings or factor.}
}
\value{
List with items \code{x} (locations converted to factor, with the factor levels telling what bars to plot) and \code{colour} (the colours to give each bar).
}
\description{
Helper function used by \code{\link[=plot_place]{plot_place()}}. Gives defaults for what bars to include in the barplot, what order to place them in, and what colour to use.
}
