% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_locationdata}
\alias{get_locationdata}
\title{Get data for locations}
\usage{
get_locationdata(x)
}
\arguments{
\item{x}{Name of the location type as string. Currently one of "forest_type", "site", "trap" or "sample".}
}
\value{
Data frame with the data of the location.
}
\description{
Helper function used by e.g. \code{\link[=get_defaults]{get_defaults()}}. Gets the data for forest types, traps or other locations.
}
\seealso{
\code{data(package="turkuwasps")} for a list of all the datasets in the package.
}
