% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{get_modelled_place}
\alias{get_modelled_place}
\title{Get the modelled number of wasps caught in each place}
\usage{
get_modelled_place(f, by, tdiff = NULL)
}
\arguments{
\item{f}{Vector of fitted values returned by \code{\link[=resample]{resample()}}, one value for each sample. Should be a vector, i.e. the matrix returned by \code{\link[=resample]{resample()}} should have every column (i.e. species) added together, or just one species selected from the matrix.}

\item{by}{Character vector giving the place (generally trap or forest type) the sample was collected. Same length as 'f'.}

\item{tdiff}{Numeric vector giving the sampling effort of each sample, in trap days. Same length as as the number of rows in 'f'. If NULL, the sampling efforts will be got from the package's internal sample data, by comparing to the row names of 'f'. If NA, results won't be scaled by sampling effort (the function will return a sampling effort of 1).}
}
\value{
List with two items:
\itemize{
\item wasps Total number of wasps in each place as estimated by the fitted model. Vector.
\item tdiff_wasps Total sampling effort in trap days of each place. Vector. If 'tdiff' was NULL, this will be a vector of ones.
}
}
\description{
Helper function used by \code{\link[=plot_modelled_place]{plot_modelled_place()}}. Gets the modelled number of wasps caught in each place, and the sampling efforts of each place if asked to scale by sampling effort.
}
\details{
Any samples which couldn't be fitted by the model (NA value in 'f') will be ignored.
}
\keyword{internal}
