% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{get_morisita}
\alias{get_morisita}
\title{Get the Morisita dissimilarity between two places}
\usage{
get_morisita(x)
}
\arguments{
\item{x}{Table of how many individuals of each species were caught in both places. Two rows (one for each place), each species in its own column.}
}
\value{
Morisita dissimilarity, ranging from 0 (no difference between the two places) to 1 (completely different places).
}
\description{
Get Morisita indexes showing how dissimilar the species assemblages of two different places were. Mainly a helper function used by function \code{\link[=morisita]{morisita()}}.
}
