% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{get_p}
\alias{get_p}
\title{Get overall p values from summary.manyglm}
\usage{
get_p(summaries, pairwise, levs)
}
\arguments{
\item{summaries}{List of the results returned by \code{\link[=summary.manyglm]{summary.manyglm()}}. The results will all be the same, except that each results compares a different level (e.g. "forest") to the other levels (e.g. "disturbed", "clearcut" etc).}

\item{pairwise}{Name of the variable to get pairwise p values for. Character string. E.g. "forest_type".}

\item{levs}{Character vector giving all the levels of the variable. (e.g. "forest", "disturbed", etc)}
}
\value{
Matrix of the p values. Square matrix with row and column names giving the two levels that were compared. (e.g. to get the p value for a difference between forest and clearcut, check row "forest" and column "clearcut")
}
\description{
Helper function used by \code{\link[=resample]{resample()}}. Extracts the p values from the results given by \code{\link[=summary.manyglm]{summary.manyglm()}}, and saves in a more convenient format. This extracts the p values between levels of a variable: e.g. if forest type has been analysed, gets the p values for differences between primary and disturbed forest, primary and clearcut etc.
}
\seealso{
\code{\link[=get_summaries]{get_summaries()}} which will typically be used to get the list given as parameter \code{summaries}.
}
\keyword{internal}
