% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{get_pairwisegroups}
\alias{get_pairwisegroups}
\title{Group together locations whose watch catches do not differ significantly}
\usage{
get_pairwisegroups(p, signif = 0.05)
}
\arguments{
\item{p}{Matrix of pairwise p values returned by \code{\link[=resample]{resample()}}, giving the p values for differences between locations.}

\item{signif}{What p value to interpret as significant. Default is for p<0.05 to be significant.}
}
\value{
List of numeric vectors. Each vector gives the indexes of locations that belong together in the same group.
}
\description{
Helper function used by \code{\link[=plot_significant_place]{plot_significant_place()}}. Groups together forest types, traps or other locations based on whether or not they differed significantly from each other in number of wasps caught.
}
\keyword{internal}
