% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{get_summaries}
\alias{get_summaries}
\title{Analyse pairwise differences with summary.manyglm}
\usage{
get_summaries(m, pairwise, model, family, analysis_args)
}
\arguments{
\item{m}{Data frame with the Malaise sample data. Must contain the columns given in \code{model}.}

\item{pairwise}{Name of the variable to get pairwise p values for. Character string. E.g. "forest_type". Must be in \code{model}.}

\item{model}{Character string or formula (see \code{\link[=as.formula]{as.formula()}}) giving the model to fit. Generally something like "mv ~ offset(tdiff_log) + days + rain + forest_type + deadwood". The variables should be found as column names in \code{m}.}

\item{family}{Probability distribution used to fit the model. Passed to \code{\link[=manyglm]{manyglm()}}. In general, this will be "negative.binomial" or "poisson".}

\item{analysis_args}{List of arguments to pass to \code{\link[=summary.manyglm]{summary.manyglm()}}.}
}
\value{
List of results given by \code{\link[=summary.manyglm]{summary.manyglm()}}. Each result will be otherwise identical, except a different level of \code{pairwise} will be compared to. E.g. first result will compare primary forest to other forest types, second result disturbed forest to other forest types, etc. One list item for each level of \code{pairwise}, except the last (since that has already been compared to all the others).
}
\description{
Helper function used by \code{\link[=resample]{resample()}}. Runs \code{\link[=summary.manyglm]{summary.manyglm()}} multiple times, to test for differences between all the levels of a categorical variable. E.g. if forest type is being analysed, analyses the differences between primary and disturbed forest, primary and clearcut, disturbed and clearcut, etc..
}
\seealso{
\code{\link[=get_p]{get_p()}} and \code{\link[=get_p_sp]{get_p_sp()}} to convert the results to something more readable.
}
\keyword{internal}
