% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Get the weights for scaling barplots}
\usage{
get_weights(tdiff, group, ...)
}
\arguments{
\item{tdiff}{Vector of sampling efforts (typically of each Malaise sample).}

\item{group}{Vector of same length as \code{tdiff}, giving the group (e.g. trap) that each sampling effort belongs to.}

\item{...}{Other arguments passed to \link{sum_by}, which passes them on to \code{\link[=aggregate]{aggregate()}}, which mostly passes them on to \code{\link[=sum]{sum()}}.}
}
\value{
Vector of the sums for each category. Named vector, groups are used as names.
}
\description{
Get how much to scale the bars of \code{\link[=plot_place]{plot_place()}} by. Typically used to count the total sampling effort of e.g. each trap, and scale down the bars by sampling effort.
}
\examples{
# get sample data (with damaged samples removed)
m = turkuwasps::malaise_sample 
m = ecology_usable(m)$samples

# get weights of each trap
get_weights(m$tdiff, m$trap) 

# get weights of each forest type
get_weights(m$tdiff, m$forest_type) 

}
