% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{include_na}
\alias{include_na}
\title{Include samples dropped by manyglm}
\usage{
include_na(f, m)
}
\arguments{
\item{f}{Results returned by \code{\link[=manyglm]{manyglm()}} when fitting a model to the data. A matrix with each sample on a separate row. Typically the fitted values.}

\item{m}{Data frame with the Malaise sample data. Must contain column "name" which gives the names of the samples.}
}
\value{
Matrix 'f' with the samples dropped by manyglm added. Same number of rows as 'm', and with the samples in the same order.
}
\description{
Helper function used by \code{\link[=resample]{resample()}}. Finds out what samples, if any, were dropped by manyglm, and returns them to the manyglm results as NA values. Manyglm will drop any samples which return NA, for example samples for which there is no rain data if rain is included in the model, which will often mess up further analyses unless fixed.
}
\keyword{internal}
