% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{interpolate_na}
\alias{interpolate_na}
\title{Get missing values by interpolating}
\usage{
interpolate_na(x)
}
\arguments{
\item{x}{Numeric vector. The first and last values must not be NA.}
}
\value{
The vector \code{x} with any missing values replaced with interpolated values.
}
\description{
Helper function used by \code{\link[=get_rarefaction]{get_rarefaction()}}. Replaces NA values in a vector with interpolated values.
}
\details{
Mainly used for rarefaction curves, to interpolate any missing values. E.g. if the first sample in the rarefaction had three wasps and one species (x=3, y=1), and the origin is (x=0, y=0), the points in between will be interpolated: (x=1, y=0.33) (x=2, y=0.66).
}
\keyword{internal}
