% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{is_datetime}
\alias{is_datetime}
\title{Check for datetimes}
\usage{
is_datetime(x)
}
\arguments{
\item{x}{Datetime or vector of datetimes.}
}
\value{
TRUE if \code{x}is a datetime, FALSE otherwise.
}
\description{
Find out if an object is a datetime object.
}
