% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{is_interval}
\alias{is_interval}
\title{Check for intervals}
\usage{
is_interval(x)
}
\arguments{
\item{x}{Interval or vector of intervals.}
}
\value{
TRUE if \code{x}is an interval, FALSE otherwise.
}
\description{
Find out if an object is an interval object.
}
