% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{is_multirow}
\alias{is_multirow}
\title{Check if a variable is a matrix with several rows}
\usage{
is_multirow(x)
}
\arguments{
\item{x}{Variable to be checked. Typically a vector or matrix}
}
\value{
TRUE if 'x' is a matrix with more than one row, otherwise FALSE. Matrixes that R calls by a different name (e.g. tables, 2D arrays etc) are counted as a matrix.
}
\description{
Helper function used by \code{\link[=plot_modelled_place]{plot_modelled_place()}}.
}
\keyword{internal}
