% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{is_na}
\alias{is_na}
\title{Check if a variable is NA (just a single NA)}
\usage{
is_na(x)
}
\arguments{
\item{x}{What to check.}
}
\value{
TRUE if \code{x} is a single NA value, otherwise FALSE.
}
\description{
Helper function. Checks if something is exactly NA. I.e. just a single NA, not e.g. a vector of NAs. Useful e.g. when \code{ìf(is.na(x))} returns an error because \code{x} may be a vector.
}
\keyword{internal}
