% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{legend_explore}
\alias{legend_explore}
\title{Add legend to extended scatterplot}
\usage{
legend_explore(breaks, col)
}
\arguments{
\item{breaks}{Numeric vector giving the breakpoints between colours. (between 0 and 1)}

\item{col}{Vector of colours, of length length(breaks)-1.}
}
\description{
Helper function used by \code{\link[=plot_explore]{plot_explore()}}. Adds a legend showing the colours used in the grid cells.
}
\keyword{internal}
