% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_rarefaction}
\alias{legend_rarefaction}
\title{Add legend to rarefaction plot}
\usage{
legend_rarefaction(
  rownames,
  colnames = NULL,
  col = "black",
  pch = NA,
  title = NULL,
  ...
)
}
\description{
No documentation yet, needs to be added.
}
