% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{length.datetime}
\alias{length.datetime}
\title{Get length of vector of datetimes}
\usage{
\method{length}{datetime}(x)
}
\arguments{
\item{x}{Vector of datetimes.}
}
\value{
Length of the vector.
}
\description{
Get the length of a vector of datetimes with \code{\link[=length]{length()}}. Datetime objects are basically a list with two items (\code{d}=datetimes, \code{tz}=time zone offsets), so unmodified \code{\link[=length]{length()}} would always return length=2, irrespective of the actual length.
}
