% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{length.interval}
\alias{length.interval}
\title{Get length of vector of intervals}
\usage{
\method{length}{interval}(x)
}
\arguments{
\item{x}{Vector of intervals.}
}
\value{
Length of the vector.
}
\description{
Get the length of a vector of intervals with \code{\link[=length]{length()}}. Interval objects are basically a list with two items (\code{s}=start datetimes, \code{e}=end datetimes), so unmodified \code{\link[=length]{length()}} would always return length=2, irrespective of the actual length.
}
