% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{match_names}
\alias{match_names}
\title{Match colour etc parameters to the right rarefaction curve}
\usage{
match_names(x, xname, levs, column_name)
}
\arguments{
\item{x}{Parameter to check. Usually a vector, e.g. for the colour of the curves: 'c(primary="darkgreen", disturbed="green", swamp="blue", clearcut="red")'.}

\item{xname}{Character string telling what parameter is being checked. E.g. "col". Used for error messages.}

\item{levs}{Character vector giving the levels of the column that is being used to split the data into separate curves (curves) or the names of parameter \code{txt} (legend texts). The curves will be drawn in this order, and the parameter's values placed in the same order. E.g. 'c("clearcut", "disturbed", "primary", "swamp")'.}

\item{column_name}{Character string telling what column is being used to split the data into separate curves. Used for error messages.}
}
\value{
Named vector of parameter values. The names tell which level each parameter value corresponds to. The vector is the same length and is in the same order as \code{levs}.
}
\description{
Helper function used by \code{\link[=plot_rarefaction]{plot_rarefaction()}} and \code{\link[=legend_rarefaction]{legend_rarefaction()}}. Checks the length and names of parameters such as \code{col} and \code{pch}. If they are named vectors, makes sure they are in the same order as the curves or legend texts that are being drawn. If not, checks the length. Only used if several curves are drawn.
}
\keyword{internal}
