% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{max.datetime}
\alias{max.datetime}
\title{Get latest datetime}
\usage{
\method{max}{datetime}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Vector of datetimes.}

\item{na.rm}{If TRUE, NA values are removed. Passed to \code{\link[=max]{max()}}.}
}
\value{
Datetime object.
}
\description{
Get the latest item in a vector of datetime objects.
}
\note{
Unlike in base \code{\link[=max]{max()}}, only one vector can be given as an argument.
}
