% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{max.interval}
\alias{max.interval}
\title{Get latest datetime in intervals}
\usage{
\method{max}{interval}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Vector of datetimes.}

\item{na.rm}{If TRUE, NA values are removed. Passed to \code{\link[=max]{max()}}.}
}
\value{
Datetime object.
}
\description{
Get the latest datetime in a vector of interval objects. Checks the start and end datetimes, and returns the latest one.
}
\note{
Unlike in base \code{\link[=max]{max()}}, only one vector can be given as an argument. Any others will be ignored.
}
