% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{means}
\alias{means}
\title{Get means of intervals}
\usage{
means(x)
}
\arguments{
\item{x}{Vector of intervals.}
}
\value{
Vector of datetimes.
}
\description{
Get the averages in between the start and end datetimes of a vector of interval objects.
}
