% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{min.datetime}
\alias{min.datetime}
\title{Get earliest datetime}
\usage{
\method{min}{datetime}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Vector of datetimes.}

\item{na.rm}{If TRUE, NA values are removed. Passed to \code{\link[=min]{min()}}.}
}
\value{
Datetime object.
}
\description{
Get the earliest item in a vector of datetime objects.
}
\note{
Unlike in base \code{\link[=min]{min()}}, only one vector can be given as an argument. Any others will be ignored.
}
