% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{min.interval}
\alias{min.interval}
\title{Get earliest datetime in intervals}
\usage{
\method{min}{interval}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Vector of datetimes.}

\item{na.rm}{If TRUE, NA values are removed. Passed to \code{\link[=min]{min()}}.}
}
\value{
Datetime object.
}
\description{
Get the earliest datetime in a vector of interval objects. Checks the start and end datetimes, and returns the earliest one.
}
\note{
Unlike in base \code{\link[=min]{min()}}, only one vector can be given as an argument. Any others will be ignored.
}
