% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morisita.R
\name{morisita}
\alias{morisita}
\title{Get the Morisita dissimilarity between places}
\usage{
morisita(x, digits = 2)
}
\arguments{
\item{x}{Table of how many individuals of each species were caught in each place. Places in rows, each species in its own column.}

\item{digits}{How many digits to round the index to. Integer, or if NA the indexes are not rounded.}
}
\value{
Matrix of Morisita dissimilarities between the places, each dissimilarity ranging from 0 (no difference between the two places) to 1 (completely different places).
}
\description{
Get Morisita indexes showing how dissimilar the species assemblages of different forest types, traps etc were. Handy way of getting a rough overview of what places have similar species in similar proportions. See Horn (1966) for a decription of the index.
}
\note{
The dissimilarities are approximate, and very small near-zero values may be off by a few hundredths because of rounding errors. The function replaces small negative values (anything between -0.04 and 0) with 0.
}
\section{References}{

\itemize{
\item Horn HS 1966. Measurement of ”Overlap” in Comparative Ecological Studies. The American Naturalist 100: 419–424. \url{http://www.jstor.org/stable/2459242}
}
}

\examples{
# get path to example wasp data
f = system.file("extdata", "wasps_example.csv", package = "turkuwasps", mustWork = TRUE)

# read the wasp data
tmp = read_wasps(f)
x = tmp$x

# get how many wasps of each species were in each forest type, and show the dissimilarities
M = table(x$forest_type, x$taxon)
morisita(M)

}
