% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nlevels0}
\alias{nlevels0}
\title{Get number of factor levels}
\usage{
nlevels0(x, ...)
}
\arguments{
\item{x}{Vector whose number of factor levels is wanted.}

\item{...}{Other arguments passed to \code{\link[=factor]{factor()}} if \code{x} is not already a factor.}
}
\value{
Number of levels of \code{x}.

seealso \code{\link[=levels0]{levels0()}}
}
\description{
Helper function used e.g. by \code{\link[=plot_place]{plot_place()}}, but can also be useful otherwise. Gets the number of factor levels of a vector. If the vector is not a factor, factors it first. Equivalent to \code{\link[=nlevels]{nlevels()}}, except handles vectors which are not a factor.
}
