% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{Ops.interval}
\alias{Ops.interval}
\title{Make basic operators work on intervals}
\usage{
\method{Ops}{interval}(e1, e2)
}
\arguments{
\item{e1}{Vector of intervals.(or numbers)}

\item{e2}{Vector of intervals (or numbers).}
}
\value{
Result of operation. Varies by operator, see details.
}
\description{
Define methods for handling some basic operations on interval objects. (e.g. \code{+}, \code{-}) This is called by R e.g. whenever numbers are added or subtracted to an interval; not meant to be called by the user.
}
\details{
Supported operators are:
\itemize{
\item \code{+} Add seconds to intervals.
\item \code{-} Subtract seconds from intervals.
\item \code{==} Check if two intervals occur at the same time.
\item \code{!=} Check if two intervals do not occur at the same time.
}
}
\seealso{
Section "Ops" in \code{?groupGeneric}, which lists all the operators.
}
