% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{overlap}
\alias{overlap}
\title{Check how much intervals overlap}
\usage{
overlap(x, with)
}
\arguments{
\item{x}{Vector of intervals to check for overlap.}

\item{with}{Interval object that the intervals in \code{x} overlap with.}
}
\value{
Vector of numbers, giving the number of seconds that each interval in \code{x} overlaps the interval with.
}
\description{
Get the number of seconds that interval objects overlap with an interval.
}
