% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{overlaps}
\alias{overlaps}
\title{Check if intervals overlap}
\usage{
overlaps(x, with)
}
\arguments{
\item{x}{Vector of intervals to check.}

\item{with}{Interval objects that \code{x} either overlap with or don't.}
}
\value{
Vector of TRUE or FALSE.
}
\description{
Check which interval objects overlap with an interval.
}
