% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{plot.datetime}
\alias{plot.datetime}
\title{Draw a scatterplot with datetimes}
\usage{
\method{plot}{datetime}(x, y, xlim = NULL, xlabel = TRUE, ...)
}
\arguments{
\item{x}{Vector of datetimes.}

\item{y}{Vector of y coordinates, same length as \code{x}.}

\item{xlim}{Interval object (see \code{\link[=as.interval]{as.interval()}}) giving the left and right limits of the plot. If not given, default limits are used.}

\item{xlabel}{If TRUE, the x axis is labelled (using default settings) by \code{\link[=axis_datetime]{axis_datetime()}}. If FALSE, the x axis is left blank.}

\item{...}{Graphical parameters and other arguments passed to \code{\link[=plot]{plot()}}, which handles the actual drawing.}
}
\value{
A list giving the x and y coordinates, returned silently. The list has four items:
\itemize{
\item \code{x} The x coordinates of each point, in plot coordinates.
\item \code{y} The y coordinates of each point, in plot coordinates.
\item \code{d} The x coordinates of each point, as datetimes.
\item \code{xlim} The left and right limits of the plot as an interval object.
}
}
\description{
Draw a scatterplot that has datetimes on the x axis. Basically a wrapper for \code{\link[=plot.default]{plot.default()}}, but converts the datetimes to plot coordinates.
}
\seealso{
\code{\link[=as.datetime]{as.datetime()}}
}
