% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_explore}
\alias{plot_explore}
\title{Plot extended scatterplot}
\usage{
plot_explore(x, y, res = 15, ...)
}
\arguments{
\item{x}{Numeric vector of the x positions of each point. E.g. the amount of rain of each sample.}

\item{y}{Numeric vector of the y positions of each point. Typically the number of wasps caught in each sample, or the number of wasps caught per day.}

\item{res}{Resolution of the grid. Default is to split the x and y axes into 15 grid lines, giving 14*14=196 grid cells.}

\item{...}{Other arguments passed to \code{\link[=image]{image()}}, which draws the grid cell colours. Will overwrite any default arguments such as \code{col}. (Argument \code{x} cannot be overwritten)}
}
\description{
Helper function used by \code{\link[=explore]{explore()}}. Plots a scatterplot of e.g. rain on the x axis and number of wasps caught in each sample on the y axis. But also adds:
\itemize{
\item A fitted line showing trends (e.g. decreasing catch the more rain there is).
\item A grid coloured in according to how many points there are in each grid cell (i.e. a 2D histogram). The counts are scaled, see "Notes".
}
}
\note{
The number of points in each grid cell is scaled before plotting, so that the colour of each grid cell shows the portion of points (of the column of cells) that are in the cell. E.g. if rain is on the x axis, then the cells of column 5 mm - 10 mm do not show how many points are in each cell; they show what portion of the points that are between rain=5mm and rain=10mm are in each cell. This makes it easier to compare different rainfall classes, even if some rainfalls have a much larger sample size (more points) than others.
}
