% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{plot_modelled_place}
\alias{plot_modelled_place}
\title{Add modelled catches to plots of wasps per place}
\usage{
plot_modelled_place(f, by, x = NULL, tdiff = NULL, plot = TRUE)
}
\arguments{
\item{f}{Matrix of fitted values returned by \code{\link[=resample]{resample()}}, one row for each sample and one column for each taxon.}

\item{by}{Character vector giving the place (generally trap or forest type) each sample was collected. Same length as the number of rows in 'f'.}

\item{x}{x coordinates to use when adding the modelled catch of each place to the plot. Can be a vector, matrix (with taxa in rows and places in columns), or a list of coordinates returned by \code{\link[=plot_place]{plot_place()}}. See 'Details' for what happens if 'x' is a matrix with more than one row (i.e. more than one taxon). If not given, default x coordinates will be used.}

\item{tdiff}{Numeric vector giving the sampling effort of each sample, in trap days. Same length as as the number of rows in 'f'. If NULL, the sampling efforts will be got from the package's internal sample data, by comparing to the row names of 'f'. If NA, results won't be scaled by sampling effort.}

\item{plot}{If TRUE, the modelled catches of each place will be added to the plot as lines + points. If FALSE, the modelled catches will only be returned, not plotted.}
}
\value{
List with two items, returned invisibly:
\itemize{
\item x Vector of x coordinates of each place.
\item y Total number of wasps for each place as estimated by the model. Scaled by sampling effort if 'tdiff' was given.
}
}
\description{
Get the modelled wasp catches per place from the fitted values returned by \code{\link[=resample]{resample()}}, and (optionally) add them to a plot. Typically used after plotting the number of wasps caught in each place with \code{\link[=plot_place]{plot_place()}}, to add a line to the plot with the model predictions.
}
\details{
If the x coordinates are given as a matrix with each species on its own row, the assumption is each species should be plotted separately. The function will count the wasp catches and x coordinates of species 1, then species 2 etc, and return wasp catch and x coodinate vectors which have the species' values pasted together (e.g. wasp catch vector has species 1 wasp catches, followed by species 2 wasp catches..). This matches how \code{\link[=plot_place]{plot_place()}} places the bars when there is more than one species plotted.
Usually the x coordinates are a vector, or a matrix with just one row. Then the species will be lumped together = the fitted values of the different species in 'f' will be combined.
Typically, 'x' will be the list returned by \code{\link[=plot_place]{plot_place()}}. Then the plotting of modelled values will automatically match the plot; the x coordinates will be extracted from the list as a matrix with either one or several rows.
}
\examples{
# get path to example wasp data
f = system.file("extdata", "wasps_example.csv", package = "turkuwasps", mustWork = TRUE)

# read the wasp data and get the corresponding sample data
tmp = read_wasps(f)
x = tmp$x
m = tmp$m

# fit model, with zero resamples since we don't need p values
model = "offset(tdiff_log) + days + rain + forest_type + deadwood"
a = resample(model, x, m, nBoot=0)

# plot how many wasps were caught in each trap versus how many the model predicts
coords = plot_place(x$trap, m=m)
plot_modelled_place(a$fit, m$trap, x=coords)
default_legend("forest_type", "Uganda 2014-2015", modelled=TRUE)

# plot how many wasps were caught and modelled in each forest type, each species separate
coords = plot_place(x$forest_type, m=m, taxon=x$taxon)
plot_modelled_place(a$fit, m$forest_type, x=coords)
default_legend("forest_type", "Uganda 2014-2015", modelled=TRUE, x="top")

}
