% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{plot_modelled_time}
\alias{plot_modelled_time}
\title{Add modelled catches to plots of wasps over time}
\usage{
plot_modelled_time(f, xlim = NULL, mdate = NULL, tdiff = NULL)
}
\arguments{
\item{f}{Matrix of fitted values returned by \code{\link[=resample]{resample()}}, one row for each sample and one column for each taxon.}

\item{xlim}{Interval object giving the x limits of the plot. Typically as returned by \code{\link[=plot_time]{plot_time()}}. If given, the modelled catches will be added as a line to the existing plot. If not, the modelled catches will only be returned, not plotted.}

\item{mdate}{Interval object giving the sample start and end dates. Same length as as the number of rows in 'f'. If NULL, the start and end dates will be got from the package's internal sample data, by comparing to the row names of 'f'.}

\item{tdiff}{Numeric vector giving the sampling effort of each sample, in trap days. Same length as as the number of rows in 'f'. If NULL, the sampling efforts will be got from the package's internal sample data, by comparing to the row names of 'f'.}
}
\value{
List with two items, returned invisibly:
\itemize{
\item x Vector of datetimes (see \code{\link[=as.datetime]{as.datetime()}}). In a sequence at one-day intervals, giving the end datetimes of the daily averages of wasp catches.
\item y Numeric vector. Daily averages of the modelled wasp catches.
}
}
\description{
Get the modelled wasp catches over time from the fitted values returned by \code{\link[=resample]{resample()}}, and (optionally) add them to a plot. Typically used after plotting the number of wasps caught each day with \code{\link[=plot_time]{plot_time()}}, to add a line to the plot with the model predictions. Currently only gives the predicted number of wasps per trap day for each day, cannot be used for plots that are not scaled down by sampling effort.
}
\examples{
# get path to example wasp data
f = system.file("extdata", "wasps_example.csv", package = "turkuwasps", mustWork = TRUE)

# read the wasp data and get the corresponding sample data
tmp = read_wasps(f)
x = tmp$x
m = tmp$m

# fit model, with zero resamples since we don't need p values
model = "offset(tdiff_log) + days + rain + forest_type + deadwood"
a = resample(model, x, m, nBoot=0)

# get the start and end of when each wasp was collected
waspdates = as.interval(x$start, x$end)

# plot how many wasps were caught each day
z = plot_time(waspdates, m, x$taxon, ylab="wasps / trap day")

# add a line showing how many wasps each day the fitted model predicts
plot_modelled_time(a$fit, xlim=z$xlim)

# add a legend
default_legend(x$taxon, x="topleft", modelled=TRUE)

}
