% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_place.R
\name{plot_place}
\alias{plot_place}
\title{Plot wasps per place}
\usage{
plot_place(x, taxon = NULL, weights = 1, defaults = TRUE, ...)
}
\arguments{
\item{x}{Vector of which trap, forest type or other location each wasp came from. Either as strings or factor. (If factor, and \code{defaults}=FALSE, one bar is plotted for each factor level, in the same order as the levels)}

\item{taxon}{Vector giving the taxon of each wasp. If given, splits the bars to show how many wasps of each taxon were caught.}

\item{weights}{How much to scale each bar by. Vector of same length as the number of bars, or can also be a single number. Typically used to scale down bars by sampling effort.}

\item{defaults}{If TRUE, uses default settings for what bars to show, and for the order and colour of the bars. For example, if \code{x} contains Ugandan traps, draws one bar for each Ugandan trap, in successional order (primary forest to farm), with primary forest in dark green, swamp in blue etc.}

\item{...}{Graphical parameters and other arguments passed to \code{\link[=barplot]{barplot()}}. These will override any default values (e.g. colours). Colours (argument \code{col}) are for each taxon if \code{taxon} was given, or for each bar if it was not.}
}
\value{
x-coordinates of the bars, returned silently (save these to variable to continue drawing on the barplot).
}
\description{
Barplot the number of wasps caught in each trap, forest type, or other location. This is basically a wrapper for two base R functions: \code{\link[=table]{table()}} for counting how many wasps belong to each bar, and \code{\link[=barplot]{barplot()}} to draw the plot.
}
