% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{plot_significant_place}
\alias{plot_significant_place}
\title{Show which bars group together in plots of wasps per place}
\usage{
plot_significant_place(p, x, signif = 0.05, plot = TRUE, ...)
}
\arguments{
\item{p}{Matrix of pairwise p values returned by \code{\link[=resample]{resample()}}, or list of such matrixes. (If a list, typically one matrix for each species.) Each square matrix gives the p values for differences between locations.}

\item{x}{Coordinates of the bars as returned by \code{\link[=plot_place]{plot_place()}}. List with elements \code{x} and \code{y}.}

\item{signif}{What p value to interpret as significant. Default is for p<0.05 to be significant.}

\item{plot}{if TRUE, the groupings will be added to the plot as lines + significance values. If FALSE, the results will ony be returned, not plotted.}

\item{...}{Graphical parameters passed to \code{\link[=segments]{segments()}} and \code{\link[=text]{text()}}. These will override any default parameters. Not all aparmeters are guaranteed to work as expected, but at least \code{cex}, \code{col}, \code{lty} and \code{lwd} can be used to adjust the text size, colour and line appearance.}
}
\value{
List with five items, returned invisibly. All are fairly complex lists, with one list item for each item in \code{p} (typically each species) and that list item split into one list item for each grouping of bars:
\itemize{
\item group Indexes of the bars that group together.
\item x0 Left x coordinates of lines grouping bars together.
\item x1 Right x coordinates of lines grouping bars together.
\item y y coordinates of lines grouping bars together. Average of the bar heights.
\item p_min Minimum pairwise p value between the bars that group together.
}
}
\description{
Group together forest types, traps or other locations based on whether or not they differed significantly from each other in number of wasps caught, and add the groupings to a plot drawn by \code{\link[=plot_place]{plot_place()}}. Draws lines onto the plot that connect bars that belong together. Also shows the smallest p value between the bars in a group, to give a rough estimate of how valid the group is.
}
\examples{
# get path to example wasp data
f = system.file("extdata", "wasps_example.csv", package = "turkuwasps", mustWork = TRUE)

# read the wasp data and get the corresponding sample data
tmp = read_wasps(f)
x = tmp$x
m = tmp$m

# fit a model and get p values (read from file by default to save time)
if (FALSE){
	model = "offset(tdiff_log) + days + rain + forest_type + deadwood"
	a = resample(model, x, m, pairwise="forest_type", nBoot=99)
} else {
	a = turkuwasps::a
}

# show how species 1 was distributed in habitat types
X = x[x$taxon == levels0(x$taxon)[1], ]
coords = plot_place(X$forest_type, m=m)

# join habitat types that were not significantly different with grey lines
plot_significant_place(a$p_pairwise_sp[1], coords)

# show the same plots for all species with more than ten individuals
i = which( table(x$taxon) > 10 )
X = x[x$taxon \%in\% levels0(x$taxon)[i], ]
coords = plot_place(X$forest_type, m=m, taxon=X$taxon, cex.names=0.5)
plot_significant_place(a$p_pairwise_sp[i], coords)

}
