% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time.R
\name{plot_time}
\alias{plot_time}
\title{Plot wasps over time}
\usage{
plot_time(
  x,
  m = NULL,
  taxon = NULL,
  xlim = default_xlims(x),
  step = 3600 * 24,
  xlabel = TRUE,
  ...
)
}
\arguments{
\item{x}{Vector of interval objects, telling when each wasp was collected.}

\item{m}{Data frame with the Malaise sample data, if wanting to scale by sampling effort (= number of traps in use at any given time). Must contain columns "start" and "end", giving start and end datetime of when each sample was collected. If NULL, the plot shows the total number of wasps collected per unit time, without taking into account how many traps were in use.}

\item{taxon}{Vector giving the taxon of each wasp. If given, shows each taxon in a different colour.}

\item{xlim}{Interval object (see \code{\link[=as.interval]{as.interval()}}) giving the desired left and right limits of the plot. If not given, default limits given by \code{\link[=default_xlims]{default_xlims()}} are used.}

\item{step}{Length in seconds of the time periods for which to count the number of wasps. Default is to plot wasps at a resolution of 86400 seconds = one day.}

\item{xlabel}{If TRUE, the x axis is labelled (using default settings) by \code{\link[=axis_datetime]{axis_datetime()}}. If FALSE, the x axis is left blank.}

\item{...}{Graphical parameters and other arguments passed to \code{\link[=barplot]{barplot()}}, which handles the actual drawing. These will override any default values (e.g. colours, which by default come from \code{\link[=default_colours]{default_colours()}}). Colours (argument \code{col}) are for each taxon, or if taxa were not given then just one colour should be given. Argument \code{xlim} is handled differently to standard \code{\link[=barplot]{barplot()}}.}
}
\value{
A list giving the x and y coordinates of each bar (= time period \code{step} seconds long), returned silently. The list has four items:
\itemize{
\item \code{x} The x coordinates of the centre of each bar, in plot coordinates. Use these to draw e.g. lines and further labels on the plot.
\item \code{y} The height of each bar, in plot coordinates.
\item \code{d} The start and end datetimes of each bar.
\item \code{xlim} The left and right limits of the plot as an interval object.
}
}
\description{
Plot the number of wasps caught each day (or other time period). This is at heart function \code{\link[=barplot]{barplot()}}, with one bar for each time period.
}
\examples{
# get example wasp data
f = system.file("extdata", "wasps_example.csv", package = "turkuwasps", mustWork = TRUE)
wasps = read_wasps(f)

# remove damaged samples and their wasps
tmp = ecology_usable(wasps)
x = tmp$wasps
m = tmp$samples

# get the start and end of when each wasp was collected
waspdates = as.interval(x$start, x$end)

# plot
plot_time(waspdates, m, x$taxon, ylab="wasps / trap day")

# plot with modified x axis
tmp = plot_time(waspdates, m, x$taxon, xlabel=FALSE)
axis_datetime(tmp$xlim, srt=270, format="\%d \%b \%Y")

}
