% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{plot_weather}
\alias{plot_weather}
\title{Plot weather data}
\usage{
plot_weather(
  what = "rain",
  xlim = NULL,
  k = 21,
  nticks = 3,
  add = FALSE,
  weather = turkuwasps::weather,
  ...
)
}
\arguments{
\item{what}{Name of the column in the weather data that should be plotted. Default is to plot rainfall.}

\item{xlim}{Interval object (see \code{\link[=as.interval]{as.interval()}}) giving the x limits of the plot. If NULL, the start and end datetimes of the Ugandan sampling are used.}

\item{k}{How much to smooth the data by (see \code{\link[=smooth_data]{smooth_data()}}). Number of days over which to smooth the data. Default (21) is to take the average of the past 21 days. If NA, the data is not smoothed.}

\item{nticks}{How many tick marks to have on the default y axis. Approximate, passed to \code{\link[=pretty]{pretty()}}.The default (3ish) prevents the axis from getting squashed in typical plotting of weather data, where there is little vertical space. If NA, the default y axis is not drawn.}

\item{add}{If TRUE, the data is added to the existing plot with \code{\link[=points.datetime]{points.datetime()}}. If FALSE, the default, draws a new plot with \code{\link[=plot.datetime]{plot.datetime()}}.}

\item{weather}{Data frame giving the weather data. Must contain columns "start", "end", and the column to be plotted. Default is to use the package's built in data \link{weather}.}

\item{...}{Graphical parameters and other arguments passed to \code{\link[=plot.datetime]{plot.datetime()}} or (if \code{add} is TRUE, to \code{\link[=points.datetime]{points.datetime()}}). These will override any default values. Typical arguments to adjust include \code{axes}, \code{bty}, \code{col}, \code{type}, \code{xlabel}.}
}
\value{
A list giving the x and y coordinates, returned silently. The list has three items:
\itemize{
\item \code{x} The datetimes of each point. These are at the end of the time period during which the weather was measured (or for smoothed data, the end of the time period over which the data was smoothed).
\item \code{y} The y coordinates of each point.
\item \code{xlim} The left and right limits of the plot as an interval object.
}
}
\description{
Plot weather data such as rainfall or temperature over time. By default, uses the data in \link{weather} and uses default colours, axes etc.
}
\examples{
# get path to example wasp data
f = system.file("extdata", "wasps_example.csv", package = "turkuwasps", mustWork = TRUE)

# read the wasp data and get the corresponding sample data
tmp = read_wasps(f)
x = tmp$x
m = tmp$m

# get x limits that cover the wasp data
xlim = plot_time(as.interval(x$start, x$end), plot=FALSE, xlabel=FALSE)$xlim

# adjust the plot layout
layout(matrix(1:2), heights=c(1, 3))

# plot rainfall
par(mar=c(1.5, 4.1, 0.5, 0.1))
plot_weather(xlim=xlim, ylab="rain (mm)")

# plot wasp catches
par(mar=c(3.5, 4.1, 2, 0.1))
plot_time(as.interval(x$start, x$end), m, x$taxon, ylab="wasps / trap day")

}
