% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{points.datetime}
\alias{points.datetime}
\title{Add points to a datetime plot}
\usage{
\method{points}{datetime}(x, y, xlim, ...)
}
\arguments{
\item{x}{Vector of datetimes.}

\item{y}{Vector of y coordinates, same length as \code{x}.}

\item{xlim}{Interval object (see \code{\link[=as.interval]{as.interval()}}) giving the left and right limits of the plot.}

\item{...}{Graphical parameters and other arguments passed to \code{\link[=points]{points()}}, which handles the actual drawing.}
}
\description{
Add points to a plot that has datetimes on the x axis. Basically a wrapper for \code{\link[=points]{points()}}, but converts the datetimes to plot coordinates.
}
\seealso{
\code{\link[=as.datetime]{as.datetime()}}
}
