% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{print.datetime}
\alias{print.datetime}
\title{Print datetimes}
\usage{
\method{print}{datetime}(x, ...)
}
\arguments{
\item{x}{Vector of datetimes.}

\item{...}{Other arguments passed to \code{\link[=print]{print()}}.}
}
\value{
Printed character vector.
}
\description{
Print datetimes in the format "2014-10-14 11:04:00 UTC+03:00". Datetime objects are basically a list with two items (\code{d}=datetimes, \code{tz}=time zone offsets), which would not display tidily with unmodified \code{\link[=print]{print()}}.
}
