% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{print.interval}
\alias{print.interval}
\title{Print intervals}
\usage{
\method{print}{interval}(x, ...)
}
\arguments{
\item{x}{Vector of intervals.}

\item{...}{Other arguments passed to \code{\link[=print]{print()}}.}
}
\value{
Printed data frame.
}
\description{
Print intervals in a readable format. Interval objects are basically a list with two items (\code{s}=start datetimes, \code{e}=end datetimes), which would not display tidily with unmodified \code{\link[=print]{print()}}.
}
