% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wasps.R
\name{read_kotka}
\alias{read_kotka}
\title{Read Kotka file}
\usage{
read_kotka(file, simplify = TRUE, columns = NULL, ...)
}
\arguments{
\item{file}{Name of the file to read from.}

\item{simplify}{If TRUE, convert the data to something more usable. Drop irrelevant columns, rename remaining columns, get forest type, trap etc from the sample data.}

\item{columns}{Named character vector giving the name of columns in the Kotka file that you want to keep if simpifying the data. For example, \code{columns=c(location="MYDocumentLocation", notes="MYNotes")} adds the columns "MYDocumentLocation" and "MYNotes" to the simplified data, renaming them to "location" and "notes". Will overwrite any standard columns of the same name. If NULL, only the standard columns "id", "sex" etc will be returned.}

\item{...}{Arguments passed to \code{\link[=read.csv]{read.csv()}}.}
}
\value{
Data frame with the wasp data. If simplified, has the following columns:
\itemize{
\item id Wasp's identifier, e.g. "ZMUT.53". (in short form, long form is e.g. "http://mus.utu.fi/ZMUT.53")
\item sex F for female, M for male, U for unknown.
\item taxon Taxon, taken from the first column with an identification. If there are several identifications, this may need overwriting.
\item event Collecting event, e.g. "Uganda 2014-2015".
\item forest_type Habitat type the wasp came from, if relevant.
\item site Site the wasp came from, if relevant.
\item trap Trap the wasp came from.
\item sample Sample the wasp came from. Usually a Malaise trap sample.
\item start Date and time when the wasp's sample started to be collected.
\item end Date and time when the wasp's sample stopped being collected.
\item tdiff Number of days that the sample was collected.
\item season Season when the sample was collected.
\item ecology_use True if the wasp can be used for ecological analyses. FALSE if its sample was damaged or otherwise unrepresentative of a normal catch. Rarely used, since \code{\link[=ecology_usable]{ecology_usable()}} will typically be used to remove unusable wasps before analyses.
}
}
\description{
Read a csv file downloaded from Kotka (Kotka Collection Management System). This is basically a wrapper for \code{\link[=read.csv]{read.csv()}}. Reads the column names ok (read.csv often corrupts them), removes extra header rows, and (optionally) simplifies the data. Simplification removes columns which are not needed, and renames other columns + adds several new columns so the data can be directly used by other functions.
}
