% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wasps.R
\name{read_wasps}
\alias{read_wasps}
\title{Read wasp data from file}
\usage{
read_wasps(file, simplify = TRUE, ...)
}
\arguments{
\item{file}{Name of the file to read from.}

\item{simplify}{If TRUE, drop irrelevant columns such as notes.}

\item{...}{Arguments passed to \link{read.csv}.}
}
\value{
Data frame with the wasp data. The following columns are added:
\itemize{
\item event
\item forest_type
\item trap
}
}
\description{
Read a csv file downloaded from Kotka (the Kotka Collection Managemen System). This is basically a wrapper for \code{\link[=read.csv]{read.csv()}}. Preserves the column names, removes extra header rows, and removes columns that are not needed. Also saves data such as the trap, forest type etc of each wasp in a format which can be directly used by other functions.
}
