% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{relevel0}
\alias{relevel0}
\title{Relevel levels of a factor or vector}
\usage{
relevel0(x, ref)
}
\arguments{
\item{x}{Vector.}

\item{ref}{Which level to move to first place. Usually character string.}
}
\value{
Vector \code{x} with the factor levels placed in a new order.
}
\description{
Helper function used e.g. by \code{\link[=resample]{resample()}}. Moves one of a vector's factor levels to first place. If the vector is not a factor, factors it first. Equivalent to \code{\link[=relevel]{relevel()}}, except handles vectors which are not a factor.
}
