% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_png.R
\name{save_png}
\alias{save_png}
\title{Save plots as PNG}
\usage{
save_png(
  f = "turkuwasp_image.png",
  width = 164,
  height = 140,
  bg = "white",
  res = 300,
  what
)
}
\arguments{
\item{f}{File path where to save the PNG file.}

\item{width}{Width of the plot in mm.}

\item{height}{Height of the plot in mm.}

\item{bg}{Background colour. Typically "transparent" or "white" (the default).}

\item{res}{Resolution as an integer. It's slightly uncertain exactly how \code{\link[=png]{png()}} handles this when using mm as units, but the bigger this number the higher the resolution.}

\item{what}{Plotting commands. The plot created by these will be saved to PNG. See examples for how to write these.}
}
\description{
Save plots to file as a PNG image. This is a wrapper for \code{\link[=png]{png()}}, with default settings for resolution, dimensions, transparency etc.
}
\examples{
\dontrun{

# save plot commands..
figure = expression( {
	barplot(1:10, space=0, col=1:10)
	legend("topleft", legend=1:10, fill=1:10)
} )

# .. then save to file
save_png("example.png", what=figure)

# alternatively, you can write the commands inside the function
save_png("example.png", what={
	barplot(1:10, space=0, col=1:10)
	legend("topleft", legend=1:10, fill=1:10)
})

# another alternative is to run as a function
figure1 = function(){
	barplot(1:10, space=0, col=1:10)
	legend("topleft", legend=1:10, fill=1:10)
}

save_png("example.png", what=figure1())

}

}
