% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{seq.datetime}
\alias{seq.datetime}
\title{Get sequence of datetimes}
\usage{
\method{seq}{datetime}(s, e, step = 3600 * 24, ...)
}
\arguments{
\item{s}{Starting datetime.}

\item{e}{End datetime.}

\item{step}{Interval between datetimes in seconds. Default is 3600*24 seconds = one day.}

\item{...}{Other arguments can be given for compatibility with \code{\link[=seq]{seq()}}, but will be ignored.}
}
\value{
Vector of datetime objects.
}
\description{
Get a sequence of datetime objects. Creates a sequence that starts from the first datetime, and stops at or after the second datetime.
}
