% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{set_tz}
\alias{set_tz}
\title{Set time zone of datetimes}
\usage{
set_tz(x, tz = "+00:00")
}
\arguments{
\item{x}{Vector of datetimes.}

\item{tz}{Character vector of time zone offsets, in format "+00:00". Can also be a single string, in which case it is recycled to the same length as \code{x}.}
}
\value{
Vector of datetime objects.
}
\description{
Set the desired time zone offset of datetime objects. This does not change the date and time, they are always stored as UTC internally. This just adjusts what time zone to display the datetimes in.
}
