% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{site}
\alias{site}
\title{Sites}
\format{
A data frame with 4 variables. The variables are:
\itemize{
\item name Name of the site, e.g. "CCT1". Mainly relevant for Uganda. For Peru, largely tells the soil type.
\item event Name of collecting event. One of "Uganda 2014-2015", "Amazon 1998", "Amazon 2000", "Amazon 2008", "Amazon 2011".
\item forest_type Name of forest type, e.g. "primary".
\item colour Colour to use by default when plotting. Different colour for each forest type.
}

Sites are in the following order: Uganda, then Peru 1998, 2000, 2008, 2011. Within each grouping they are sorted by successional status then alphabetical order (e.g. Uganda CC to FARM).
}
\usage{
site
}
\description{
Ugandan and Peruvian sites. Used by some functions to get all the sites, what order to plot them in by default, and what colours to use.
}
\keyword{datasets}
