% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{smooth_data}
\alias{smooth_data}
\title{Smooth data that has been collected during datetime intervals}
\usage{
smooth_data(x, xdate, k = 28, p = 0.25, xlim = NULL, step = 3600 * 24)
}
\arguments{
\item{x}{Numeric vector of data values, such as rainfall per day, or temperature.}

\item{xdate}{Vector of intervals (see \code{\link[=as.interval]{as.interval()}}) of same length as \code{x}, giving the times when the data was collected.}

\item{k}{Number of days over which to smooth the data. Default (28) is to take the average of the past 28 days. Larger numbers give smoother results.}

\item{p}{Minimum portion of the data that is required. Filters out random spikes at the start and end of when there is data. Default is to return NA if less than 25\% of the period over which the data is averaged has data.}

\item{xlim}{Interval object giving the start and end of the time period for which smoothed data is wanted. If not given, the smoothed data will cover the entire period from which there is data.}

\item{step}{Desired resolution (in seconds) of the smoothed data. Default is to return a smoothed data point every 24 hours.}
}
\value{
List giving the datetimes and values of the smoothed data. The list has two items:
\itemize{
\item \code{x} Vector of datetimes giving the end of the datat point's time interval.
\item \code{y} Vector of smoothed data.
}
}
\description{
Smooth e.g. weather data that has been collected at specific time intervals. Useful to e.g. show average rainfall.
}
\details{
The default smoothing takes the average of the past 28 days' data, at one-day intervals. Data that was collected at the start or end of the 28 day period is weighted down: if e.g. only half of the time period during which the data was collected is in the 28-day period, it'll have a 50\% weight in the weighted average.
}
\examples{
# get the start and end times of weather data
wdate = as.interval(weather$start, weather$end)

# get the amount of rain per day
rain = weather$rain / tdiff(wdate) * 3600 * 24

# smooth the rain data
r = smooth_data(rain, wdate)

# plot the raw data as points, and the smoothed data as a line
xlim = as.interval(min(wdate), max(wdate))
plot(wdate$e, rain, xlim=xlim, pch=20, cex=0.1, col="blue")
lines(r$x, r$y, xlim=xlim, col="blue")

}
