% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{[.datetime}
\alias{[.datetime}
\title{Select items of a datetime vector}
\usage{
\method{[}{datetime}(x, i = 1:length(x))
}
\arguments{
\item{x}{Vector of datetimes.}

\item{i}{Indexes of items to select. Default is to select everything.}
}
\value{
Vector of datetimes.
}
\description{
Called by R whenever items are selected by index in a vector of datetime objects (e.g. \code{x[1]}). Not meant to be called by the user. Datetime objects are basically a list with two items (\code{d}=datetimes, \code{tz}=time zone offsets), not vectors, so the standard square brackets would not work properly.
}
