% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{[.interval}
\alias{[.interval}
\title{Select items of an interval vector}
\usage{
\method{[}{interval}(x, i = 1:length(x))
}
\arguments{
\item{x}{Vector of intervals.}

\item{i}{Indexes of items to select. Default is to select everything.}
}
\value{
Vector of intervals.
}
\description{
Called by R whenever items are selected by index in an interval of datetime objects (e.g. \code{x[1]}). Not meant to be called by the user. Interval objects are basically a list with two items (\code{s}=start datetimes, \code{e}=end datetimes), not vectors, so the standard square brackets would not work properly.
}
