% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{[<-.datetime}
\alias{[<-.datetime}
\title{Save datetimes into a vector}
\usage{
\method{[}{datetime}(x, i, j = "not used") <- value
}
\arguments{
\item{x}{Vector of datetimes.}

\item{i}{Indexes of items where to save.}

\item{j}{Parameter called by R, but I'm not sure what, if anything, it does. Not used.}

\item{value}{The datetime(s) to be saved in \code{x}.}
}
\value{
Vector of datetimes.
}
\description{
Called by R whenever datetime objects are saved by index into a vector of datetimes (e.g. \code{x[2] = d}). Not meant to be called by the user. Datetime objects are basically a list with two items (\code{d}=datetimes, \code{tz}=time zone offsets), not actually vectors, so the standard code for saving would not work properly.
}
