% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{[<-.interval}
\alias{[<-.interval}
\title{Save intervals into a vector}
\usage{
\method{[}{interval}(x, i, j = "not used") <- value
}
\arguments{
\item{x}{Vector of intervals.}

\item{i}{Indexes of items where to save.}

\item{j}{Parameter called by R, but I'm not sure what, if anything, it does. Not used.}

\item{value}{interval(s) to be saved in \code{x}.}
}
\value{
Vector of intervals.
}
\description{
Called by R whenever interval objects are saved by index into a vector of intervals (e.g. \code{x[2] = d}). Not meant to be called by the user. Interval objects are basically a list with two items (\code{s}=start datetimes, \code{e}=end datetimes), not actually vectors, so the standard code for saving would not work properly.
}
