% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sum_by}
\alias{sum_by}
\title{Add together all the numbers that belong to the same category}
\usage{
sum_by(x, by, ...)
}
\arguments{
\item{x}{Vector of numbers to add together.}

\item{by}{Vector of same length as \code{x}, giving the category that each number belongs to}

\item{...}{Other arguments passed to \code{\link[=aggregate]{aggregate()}}, which mostly passes them on to \code{\link[=sum]{sum()}}.}
}
\value{
Vector of the sums for each category. Named vector, categories are used as names.
}
\description{
Split a numeric vector into categories and get the sum of each category. This is basically a more readable wrapper for \code{\link[=aggregate]{aggregate()}}.
}
\examples{
# save two vectors giving the number of wasps caught in different places, and the place names
n = c(1, 2, 3, 4, 5, 6)
place = c("forest", "swamp", "farm", "forest", "swamp", "farm")

# get the total number of wasps in each place
sum_by(n, by=place)

}
