% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{tdiff}
\alias{tdiff}
\title{Get length of interval}
\usage{
tdiff(x)
}
\arguments{
\item{x}{Vector of intervals.}
}
\value{
Numeric vector.
}
\description{
Find out how many seconds there are between the start and end of an interval object.
}
