% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{wasps_example}
\alias{wasps_example}
\title{Example wasp data}
\format{
A csv file with 456 data rows (each corresponding to one wasp) and 58 variables in columns. There are also two header rows. Kotka variables are documented in https://kotka.luomus.fi/documentation/field (link valid 2023-04-20).
}
\description{
Example csv file of wasp data downloaded from Kotka (Kotka Collection Management System). Ugandan rhyssine wasps, as published in https://doi.org/10.1098/rsos.190913 and https://doi.org/10.3897/zookeys.878.37845. Used in the examples to show what wasp data typically looks like.
}
\seealso{
\code{\link[=read_wasps]{read_wasps()}} which reads data from files like this and converts to something easier to read and use.
}
