% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{weather}
\alias{weather}
\title{Weather}
\format{
A data frame with 8 variables. The variables are:
\itemize{
\item event Name of collecting event. One of "Uganda 2014-2015", "Amazon 1998", "Amazon 2000", "Amazon 2008", "Amazon 2011".
\item start Datetime when weather started to be measured, e.g. "2014-10-14 09:00:00 UTC+03:00". There is some guesswork involved in these: I believe that Ugandan rain, min and max temperature were collected mornings about 9 AM, and are for the previous 24h; and that in Peru they are for the period 8 AM to next day 8 AM.
\item end Datetime when weather stopped being measured, e.g. "2014-10-15 09:00:00 UTC+03:00". There is some guesswork involved in these: I believe that Ugandan rain, min and max temperature were collected mornings about 9 AM, and are for the previous 24h; and that in Peru they are for the period 8 AM to next day 8 AM.
\item rain Amount of rain in millimetres during the (24 hour) measurement interval.
\item t_min Minimum temperature in Celsius during the (24 hour) measurement interval. For the Ugandan data, this is from the forest thermometer.
\item t_max Maximum temperature in Celsius during the (24 hour) measurement interval. For the Ugandana data, this si from the forest thermometer.
\item t_avg Average temperature in Celsius during the (24 hour) measurement interval. Only for Peruvian weather data.
\item season One of "dry" or "wet", giving an approximate estimate of whether the measurement occurred during the dry or wet season. Only for Ugandan weather data.
}
}
\usage{
weather
}
\description{
Weather data from Uganda and Peru. Used e.g. to plot rainfall or temperature.
}
\seealso{
Zenodo has more details on the Ugandan and Peruvian weather data:
https://doi.org/10.5281/zenodo.2225643
https://doi.org/10.5281/zenodo.4030168
}
\keyword{datasets}
