% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{as.character.datetime}
\alias{as.character.datetime}
\title{Convert datetime to character}
\usage{
\method{as.character}{datetime}(x)
}
\arguments{
\item{x}{Vector of datetimes.}
}
\value{
Character vector.
}
\description{
Convert datetime objects to strings of the format "2014-10-14 11:04:00 UTC+03:00".
}
