% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{as_italic}
\alias{as_italic}
\title{Print text as italic}
\usage{
as_italic(x)
}
\arguments{
\item{x}{Character vector which is to be converted to italic.}
}
\value{
Vector which when displayed in a plot, is in italics.
}
\description{
Helper function used by \code{\link[=default_legend]{default_legend()}}. Makes a character vector (typically species names) display in italic font. Exactly how this works is a bit of a mystery to me: R handles italics, bold etc rather unclearly.
}
