% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{draw_rarefaction}
\alias{draw_rarefaction}
\title{Draw rarefaction curves}
\usage{
draw_rarefaction(r, ci = FALSE, add = FALSE, pch = NULL, pch_col = NULL, ...)
}
\description{
No documentation yet, needs to be added, helper function
... passed to plot and lines
}
