% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wasps.R
\name{ecology_usable}
\alias{ecology_usable}
\title{Get data that is usable in ecological analyses}
\usage{
ecology_usable(x, unusable = "")
}
\arguments{
\item{x}{Data frame with the wasp data. Must contain columns "sample" and "trap", which give the Malaise sample and trap that each wasp came from.}

\item{unusable}{Character vector giving any additional samples that are unusable. Case sensitive (e.g. "CCT1-141022", not "cct1-141022").}
}
\value{
List with four elements:
\itemize{
\item \code{wasps} The wasp data without wasps that came from unusable samples.
\item \code{samples} The sample data (i.e. data frame \link{malaise_sample}) without unusable samples.
\item \code{removed_wasps} The row numbers of wasps which were removed as unusable. Useful if wanting to go back to the raw data, to see which wasps were kept and which filtered out.
\item \code{removed_samples} The row numbers of samples which were removed as unusable. Useful if wanting to go back to the raw data, to see which samples were kept and which filtered out.
}
}
\description{
Get the wasp and sample data that can be used in ecological analyses. Removes samples that were damaged in the field, and any other samples marked by the user as being unusable (typically because the wasp jar rotted). Removes wasps from these samples, and also any wasps that have no data on what trap they came from.
}
