% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecology_usable.R
\name{ecology_usable}
\alias{ecology_usable}
\title{Get data that is usable in ecological analyses}
\usage{
ecology_usable(x, unusable = "")
}
\arguments{
\item{x}{Data frame with the wasp data. Must contain columns "sample" and "trap", which give the Malaise sample and trap that each wasp came from.}

\item{unusable}{Character vector giving any additional samples that are unusable. Case sensitive (e.g. "CCT1-141022", not "cct1-141022").}
}
\value{
List with two elements:
\itemize{
\item \code{wasps} The wasp data without wasps that came from unusable samples.
\item \code{samples} The sample data (i.e. data frame \link{malaise_sample}) without unusable samples.
}
}
\description{
Get the wasp and sample data that can be used in ecological analyses. Removes samples that were damaged in the field, and any other samples marked by the user as being unusable (typically because the wasp jar rotted). Removes wasps from these samples, and also any wasps that have no data on what trap they came from. Also removes any samples that came from other collecting events than the event(s) the wasps came from.
}
