% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_place.R
\name{get_weights}
\alias{get_weights}
\title{Get the weights for scaling barplots}
\usage{
get_weights(barnames, m)
}
\arguments{
\item{barnames}{Vector of bar names for which the weights are desired. Used to find out if the bars are forest types, sites or traps, and to give the correct weights in the correct order for each bar. Must be valid forest type, site or trap names. Either as strings or factor.}

\item{m}{Data frame with the Malaise sample data. Used to get the sampling effort of each location. Must contain columns "tdiff" (sampling effort) and one of "forest_type", "site" or "trap" (whatever is being plotted).}
}
\value{
Vector of weights with which to multiply the bar heights. Named vector, in same order as the bars.
}
\description{
Helper function for scaling the bars of \code{\link[=plot_place]{plot_place()}}. Counts the total sampling effort of each bar (e.g. each trap), and tells how to scale the bars by sampling effort.
}
\keyword{internal}
