% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{levels0}
\alias{levels0}
\title{Get factor levels}
\usage{
levels0(x, ...)
}
\arguments{
\item{x}{Vector whose factor levels are wanted.}

\item{...}{Other arguments passed to \code{\link[=factor]{factor()}} if \code{x} is not already a factor.}
}
\value{
Character vector giving the levels of \code{x}.

seealso \code{\link[=nlevels0]{nlevels0()}}, \code{\link[=relevel0]{relevel0()}}
}
\description{
Get the factor levels of a vector. If the vector is not a factor, factor it first (i.e. equivalent to calling \code{levels(factor(x))}). Basically a more readable wrapper for \code{\link[=levels]{levels()}} and \code{\link[=factor]{factor()}}.
}
\examples{
# 'x' contains three different forest types: primary, disturbed, or farm
x = c("primary", "disturbed", "primary", "primary", "farm")
levels0(x)

}
