% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_package_and_data.R
\docType{data}
\name{malaise_sample}
\alias{malaise_sample}
\title{Samples}
\format{
A data frame with 10 variables. The variables are:
\itemize{
\item name Name of the sample, e.g. "CCT1-141022".
\item event Name of collecting event. One of "Uganda 2014-2015", "Amazon 1998", "Amazon 2000", "Amazon 2008", "Amazon 2011".
\item forest_type Name of forest type, e.g. "primary".
\item site Name of site, e.g. "CC". Mainly relevant for Uganda. For Peru, largely tells the soil type.
\item trap Name of trap, e.g. "CCT1".
\item colour Colour to use by default when plotting. Different colour for each forest type.
\item latitude Latitude of trap, in decimal degrees.
\item start Datetime when sample started to be collected by the Malaise trap, e.g. "2014-10-14 11:04:00 UTC+03:00".
\item end Datetime when sample stopped being collected by the Malaise trap, e.g. "2014-10-22 11:34:00 UTC+03:00".
\item tdiff Length of time the sample was collected, in days.
\item damaged TRUE if the sample was damaged in some way (e.g. trampled by elephant), or doesn't have a complete catch for some other reason. FALSE if all is ok. Damaged samples and wasps from them should not be included in ecological analyses.
}

Samples are in the following order: Uganda, then Peru 1998, 2000, 2008, 2011. Within each grouping they are sorted by successional status then alphabetical order (e.g. Uganda primary forest samples to farm samples).
}
\usage{
malaise_sample
}
\description{
Ugandan and Peruvian Malaise samples. Used by e.g. \code{\link[=read_wasps]{read_wasps()}} to get forest type etc for each wasp, and by other functions to find out when and how long each sample was collected. Could also be used to get all the samples, what order to plot them in by default, and what colours to use.
}
\keyword{datasets}
