% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{plot_rarefaction}
\alias{plot_rarefaction}
\title{Plot rarefaction curves}
\usage{
plot_rarefaction(
  x,
  n = 10,
  p = 0.84,
  by = NULL,
  ci = FALSE,
  add = FALSE,
  pch = NULL,
  ...
)
}
\description{
No documentation yet, needs to be added, x must have columns taxon and sample
... passed to plot and lines
}
