% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wasps.R
\name{read_wasps}
\alias{read_wasps}
\title{Read wasp data from file}
\usage{
read_wasps(file, simplify = TRUE, ...)
}
\arguments{
\item{file}{Name of the file to read from.}

\item{simplify}{If TRUE, drop irrelevant columns such as notes.}

\item{...}{Arguments passed to \link{read.csv}.}
}
\value{
Data frame with the wasp data. If simplified, has the following columns:
\itemize{
\item id
\item event
\item forest_type
\item trap
\item sample
}
}
\description{
Read a csv file downloaded from Kotka (Kotka Collection Managemen System). This is basically a wrapper for \code{\link[=read.csv]{read.csv()}}. Preserves the column names, removes extra header rows, and (optionally) simplifies the columns. Simplification removes columns which are not needed, and renames other columns + adds several new columns so the data can be directly used by other functions.
}
