% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{relevel0}
\alias{relevel0}
\title{Relevel levels of a factor or vector}
\usage{
relevel0(x, ref)
}
\arguments{
\item{x}{Vector.}

\item{ref}{Which level to move to first place. Usually character string.}
}
\value{
Vector \code{x} with the factor levels placed in a new order.

seealso \code{\link[=nlevels0]{nlevels0()}}, \code{\link[=levels0]{levels0()}}
}
\description{
Move one of a vector's factor levels to first place. If the vector is not a factor, factor it first. Equivalent to \code{\link[=relevel]{relevel()}}, except handles vectors which are not a factor.
}
\examples{
# 'x' contains three different forest types: primary, disturbed, or farm
x = c("primary", "disturbed", "primary", "primary", "farm")
levels0(x)

# reorder the levels so that primary is shown first
x = relevel0(x, "primary")
levels0(x)

}
