% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Fit glm and analyse it}
\usage{
resample(model, x, m, pairwise = NULL, family = "negative.binomial", ...)
}
\arguments{
\item{model}{Character string giving the model to fit. Generally something like "offset(tdiff_log) + days + rain + forest_type + deadwood". The variables should be found as column names in \code{m}. Converted to \link{formula}, and will accept e.g. interaction terms in the same format as for formulae.}

\item{x}{Data frame with the wasp data. Must contain columns "sample" and "taxon".}

\item{m}{Data frame with the Malaise sample data. Must contain columns "name" and "event".}

\item{pairwise}{Character vector of the columns in \code{m} for which pairwise p values should be calculated. E.g. \code{pairwise = "forest_type"} will check what forest types differ significantly from each other, not just if significant differences exist between forest types. Should only be used for categorical variables such as forest type, trap etc; not numeric such as rainfall. If NULL, only checks if significant differences exist.}

\item{family}{Probability distribution used to fit the model. Passed to \code{\link[=manyglm]{manyglm()}}. In general, this will be "negative.binomial" or "poisson".}

\item{...}{Other parameters passed to \code{\link[=anova.manyglm]{anova.manyglm()}}, and \code{\link[=summary.manyglm]{summary.manyglm()}}. These will override any default parameters. In general, there will be little need to adjust anything except \code{nBoot} (number of resamples). It is nine by default, to give a quick but approximate result. More accurate results will need e.g. \code{nBoot=499}, which can take several hours to finish.}
}
\value{
List with items:
\itemize{
\item anova Results of \code{\link[=anova.manyglm]{anova.manyglm()}}
\item fit Fitted values for each sample and taxon.
}
}
\description{
Fit generalised linear models to the wasp catches, and get the p values for whether forest type, rain etc affect the catch of each species. This is basically a wrapper for the package \link{mvabund}, functions \code{\link[=manyglm]{manyglm()}}, \code{\link[=anova.manyglm]{anova.manyglm()}}, and \code{\link[=summary.manyglm]{summary.manyglm()}}. The analyses take a long time to finish, anything from some minutes (\code{nBoot=9}) to hours (\code{nBoot=499})!
}
