% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_png.R
\name{save_png}
\alias{save_png}
\title{Save plots as PNG}
\usage{
save_png(f = "turkuwasp_image.png", what)
}
\arguments{
\item{f}{File path where to save the PNG file.}

\item{what}{Plotting commands. The plot created by these will be saved to PNG. See examples for how to write these.}
}
\description{
Save plots to file as a PNG image. This is a wrapper for \code{\link[=png]{png()}}, with default settings for resolution, dimensions, transparency etc.
}
\examples{
\dontrun{

save_png("example.png", {
	barplot(1:10, space=0, col=1:10)
	legend("topleft", legend=1:10, fill=1:10)
})

}

}
