% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sum_by}
\alias{sum_by}
\title{Add together all the numbers that belong to the same category}
\usage{
sum_by(x, by, ...)
}
\arguments{
\item{x}{Vector of numbers to add together.}

\item{by}{Vector of same length as \code{x}, giving the category that each number belongs to}

\item{...}{Other arguments passed to \code{\link[=aggregate]{aggregate()}}, which mostly passes them on to \code{\link[=sum]{sum()}}.}
}
\value{
Vector of the sums for each category. Named vector, categories are used as names.
}
\description{
Helper function used by \code{\link[=get_weights]{get_weights()}}. Used to count the total sampling effort for each forest type, trap or other location. This is basically a wrapper for \code{\link[=aggregate]{aggregate()}}.
}
